add product: add test for prod_type route
